skip test ports with flat memory for test_get_transceiver_bulk_status
