+++ tests/platform_tests/mellanox/util.py
@@ -1,191 +1,6 @@
 import re
 import logging
 import ast
-from tests.common.platform.transceiver_utils import get_map_port_to_start_and_end_line_number_for_sfp_eeporm
-
-pattern_top_layer_key_value = r"^(?P<key>Ethernet\d+):(?P<value>.*)"
-pattern_second_layer_key_value = r"(^\s{8}|\t{1})(?P<key>[a-zA-Z0-9][a-zA-Z0-9\s\/\(\)-]+):(?P<value>.*)"
-pattern_third_layer_key_value = r"(^\s{16}|\t{2})(?P<key>[a-zA-Z0-9][a-zA-Z0-9\s\/]+):(?P<value>.*)"
-
-pattern_digit_unit = r"^(?P<digit>-[0-9\.]+|[0-9.]+)(?P<unit>dBm|mA|C|c|Volts)"
-
-
-def parse_one_sfp_eeprom_info(sfp_eeprom_info, start, end):
-    """
-    Parse the one sfp eeprom info, return top_key, sfp_eeprom_info_dict
-    e.g
-    sfp_info:
-    Ethernet0: SFP EEPROM detected
-        Application Advertisement: N/A
-        Connector: No separable connector
-        Encoding: 64B/66B
-        Extended Identifier: Power Class 3 Module (2.5W max.),
-                             No CLEI code present in Page 02h, CDR present in TX, CDR present in RX
-        Extended RateSelect Compliance: Unknown
-        Identifier: QSFP28 or later
-        Length Cable Assembly(m): 3.0
-        Nominal Bit Rate(100Mbs): 255
-        Specification compliance:
-                10/40G Ethernet Compliance Code: Extended
-                Extended Specification Compliance: 100G AOC (Active Optical Cable) or 25GAUI C2M AOC
-                Fibre Channel Link Length: Unknown
-                Fibre Channel Speed: Unknown
-                Fibre Channel Transmission Media: Unknown
-                Fibre Channel Transmitter Technology: Unknown
-                Gigabit Ethernet Compliant Codes: Unknown
-                SAS/SATA Compliance Codes: Unknown
-                SONET Compliance Codes: Unknown
-        Vendor Date Code(YYYY-MM-DD Lot): 2019-01-17
-        Vendor Name: Mellanox
-        Vendor OUI: 00-02-c9
-        Vendor PN: MFA1A00-C003
-        Vendor Rev: B2
-        Vendor SN: MT1903FT05965
-        ChannelMonitorValues:
-                RX1Power: 0.927dBm
-                RX2Power: 0.938dBm
-                RX3Power: 0.912dBm
-                RX4Power: 0.95dBm
-                TX1Bias: 6.75mA
-                TX1Power: 1.071dBm
-                TX2Bias: 6.75mA
-                TX2Power: 1.04dBm
-                TX3Bias: 6.75mA
-                TX3Power: 1.039dBm
-                TX4Bias: 6.75mA
-                TX4Power: 1.031dBm
-        ChannelThresholdValues:
-                RxPowerHighAlarm  : 5.4dBm
-                RxPowerHighWarning: 2.4dBm
-                RxPowerLowAlarm   : -13.307dBm
-                RxPowerLowWarning : -10.301dBm
-                TxBiasHighAlarm   : 8.5mA
-                TxBiasHighWarning : 8.0mA
-                TxBiasLowAlarm    : 5.492mA
-                TxBiasLowWarning  : 6.0mA
-        ModuleMonitorValues:
-                Temperature: 43.105C
-                Vcc: 3.235Volts
-        ModuleThresholdValues:
-                TempHighAlarm  : 80.0C
-                TempHighWarning: 70.0C
-                TempLowAlarm   : -10.0C
-                TempLowWarning : 0.0C
-                VccHighAlarm   : 3.5Volts
-                VccHighWarning : 3.465Volts
-                VccLowAlarm    : 3.1Volts
-                VccLowWarning  : 3.135Volts
-    top_key, sfp_eeprom_info_dict:
-    Ethernet0,
-    {
-        'Ethernet0': 'SFP EEPROM detected',
-        'Application Advertisement': 'N/A',
-        'Connector': 'No separable connector',
-        'Encoding': '64B/66B',
-        'Extended Identifier': 'Power Class 3 Module (2.5W max.),
-                               No CLEI code present in Page 02h, CDR present in TX, CDR present in RX',
-        'Extended RateSelect Compliance': 'Unknown',
-        'Identifier': 'QSFP28 or later',
-        'Length Cable Assembly(m)': '3.0',
-        'Nominal Bit Rate(100Mbs)': '255',
-        'Specification compliance': {
-            '10/40G Ethernet Compliance Code': 'Extended',
-            'Extended Specification Compliance': '100G AOC (Active Optical Cable) or 25GAUI C2M AOC',
-            'Fibre Channel Link Length': 'Unknown',
-            'Fibre Channel Speed': 'Unknown',
-            'Fibre Channel Transmission Media': 'Unknown',
-            'Fibre Channel Transmitter Technology': 'Unknown',
-            'Gigabit Ethernet Compliant Codes': 'Unknown',
-            'SAS/SATA Compliance Codes': 'Unknown',
-            'SONET Compliance Codes': 'Unknown'
-        },
-        'Vendor Date Code(YYYY-MM-DD Lot)': '2019-01-17',
-        'Vendor Name': 'Mellanox',
-        'Vendor OUI': '00-02-c9',
-        'Vendor PN': 'MFA1A00-C003',
-        'Vendor Rev': 'B2',
-        'Vendor SN': 'MT1903FT05965',
-        'ChannelMonitorValues': {
-            'RX1Power': '0.927dBm',
-            'RX2Power': '0.938dBm',
-            'RX3Power': '0.912dBm',
-            'RX4Power': '0.95dBm',
-            'TX1Bias': '6.75mA',
-            'TX1Power': '1.071dBm',
-            'TX2Bias': '6.75mA',
-            'TX2Power': '1.04dBm',
-            'TX3Bias': '6.75mA',
-            'TX3Power': '1.039dBm',
-            'TX4Bias': '6.75mA',
-            'TX4Power': '1.031dBm'
-        },
-        'ChannelThresholdValues': {
-            'RxPowerHighAlarm': '5.4dBm',
-            'RxPowerHighWarning': '2.4dBm',
-            'RxPowerLowAlarm': '-13.307dBm',
-            'RxPowerLowWarning': '-10.301dBm',
-            'TxBiasHighAlarm': '8.5mA',
-            'TxBiasHighWarning': '8.0mA',
-            'TxBiasLowAlarm': '5.492mA',
-            'TxBiasLowWarning': '6.0mA'
-        },
-        'ModuleMonitorValues': {
-            'Temperature': '43.105C',
-            'Vcc': '3.235Volts'
-        },
-        'ModuleThresholdValues': {
-            'TempHighAlarm': '80.0C',
-            'TempHighWarning': '70.0C',
-            'TempLowAlarm': '-10.0C',
-            'TempLowWarning': '0.0C',
-            'VccHighAlarm': '3.5Volts',
-            'VccHighWarning': '3.465Volts',
-            'VccLowAlarm': '3.1Volts',
-            'VccLowWarning': '3.135Volts'
-        }
-    }
-    """
-    one_sfp_eeprom_info_dict = {}
-    second_layer_dict = {}
-    previous_key = ""
-    top_key = ""
-    for line in sfp_eeprom_info[start:end+1]:
-        res1 = re.match(pattern_top_layer_key_value, line)
-        if res1:
-            top_key = res1.groupdict()["key"].strip()
-            one_sfp_eeprom_info_dict[top_key] = res1.groupdict()[
-                "value"].strip()
-            continue
-        res2 = re.match(pattern_second_layer_key_value, line)
-        if res2:
-            if second_layer_dict and previous_key:
-                one_sfp_eeprom_info_dict[previous_key] = second_layer_dict
-                second_layer_dict = {}
-            one_sfp_eeprom_info_dict[res2.groupdict()["key"].strip()] = res2.groupdict()[
-                "value"].strip()
-            previous_key = res2.groupdict()["key"].strip()
-        else:
-            res3 = re.match(pattern_third_layer_key_value, line)
-            if res3:
-                second_layer_dict[res3.groupdict()["key"].strip()] = res3.groupdict()[
-                    "value"].strip()
-    if second_layer_dict and previous_key:
-        one_sfp_eeprom_info_dict[previous_key] = second_layer_dict
-
-    return top_key, one_sfp_eeprom_info_dict
-
-
-def parse_sfp_eeprom_infos(eeprom_infos):
-    """
-    This method is to pares sfp eeprom infos, and return sfp_eeprom_info_dict
-    """
-    map_port_to_start_and_end_line_nubmer, sfp_eeprom_list = get_map_port_to_start_and_end_line_number_for_sfp_eeporm(
-        eeprom_infos)
-    sfp_eeprom_info_dict = {}
-    for port_name, line_number_section in map_port_to_start_and_end_line_nubmer.items():
-        _, eeprom_info = parse_one_sfp_eeprom_info(sfp_eeprom_list, line_number_section[0], line_number_section[1])
-        sfp_eeprom_info_dict[port_name] = eeprom_info
-    return sfp_eeprom_info_dict
 
 
 def check_sfp_eeprom_info(duthost, sfp_eeprom_info, is_support_dom, show_eeprom_cmd, is_flat_memory):
